Only let the path stream know of the memory manager for easy materializing
