+++ src/main/java/com/github/giedomak/telepathdb/physicaloperators/OpenHashJoin.kt
@@ -7,18 +7,22 @@
 
 package com.github.giedomak.telepathdb.physicaloperators
 
+import com.github.giedomak.telepathdb.TelepathDB
 import com.github.giedomak.telepathdb.datamodels.graph.PathStream
-import com.github.giedomak.telepathdb.datamodels.stores.PathIdentifierStore
-import com.github.giedomak.telepathdb.memorymanager.MemoryManager
 import com.github.giedomak.telepathdb.utilities.Logger
+import java.util.*
 
 /**
  * Open hash-join to maintain compatibility with concatenating streams instead of evaluating physical plans.
  */
-class OpenHashJoin(private val stream1: PathStream, private val stream2: PathStream, private val pathIdentifierStore: PathIdentifierStore = PathIdentifierStore) {
+class OpenHashJoin(
+        private val stream1: PathStream,
+        private val stream2: PathStream,
+        private val telepathDB: TelepathDB = TelepathDB
+) {
 
     /**
-     * Join two streams of Paths following the HashJoin algorithm and by using our MemoryManager.
+     * Join two streams of Paths following the HashJoin algorithm and by using our SimpleMemoryManager.
      *
      * @param stream1 First stream of paths we'll join on its lastNode().
      * @param stream2 Second stream of paths we'll join on its firstNode().
@@ -26,19 +30,19 @@ class OpenHashJoin(private val stream1: PathStream, private val stream2: PathStr
      */
     fun evaluate(): PathStream {
 
-        // Make sure we get a free slot in the MemoryManager
-        val offset = MemoryManager.maxId + 1
+        // TODO: Make sure the memoryManager can guarantee a free spot
+        val salt = Random().nextLong()
 
         // Put all Paths from stream1 into a HashMap with the lastNode() as key
-        stream1.paths.forEach { MemoryManager[offset + it.nodes.last().id] = it }
+        stream1.paths.forEach { telepathDB.memoryManager[salt + it.nodes.last().hashCode().toLong()] = it }
 
         Logger.debug("Done creating the hashTable, now concatenating")
 
         // Get all Paths from the HashMap which have the firstNode() as key, and concatenate
         return PathStream(
                 stream2.paths.flatMap { v2 ->
-                    MemoryManager[offset + v2.nodes.first().id]
-                            .map { v1 -> pathIdentifierStore.concatenatePaths(v1, v2) }
+                    telepathDB.memoryManager[salt + v2.nodes.first().hashCode().toLong()]
+                            .map { v1 -> telepathDB.pathIdentifierStore.concatenatePaths(v1, v2) }
                 }
         )
     }
